Move `EquirectUVNode` to `equirectUV()` TSL function
